Refactor: CarGroup을 삭제하고 RacingGame으로 역할 이전
