+++ src/main/java/domain/RacingGame.java
@@ -3,25 +3,55 @@ package domain;
 import utils.PowerGenerator;
 import utils.RandomPowerGenerator;
 
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
 public class RacingGame {
-    private final CarGroup carGroup;
+    private final List<Car> carGroup;
     private final PowerGenerator powerGenerator;
 
 
-    public RacingGame(final CarGroup carGroup) {
-        this.carGroup = carGroup;
+    public RacingGame(final String[] carNames) {
+        this.carGroup = makeCarGroup(carNames);
         this.powerGenerator = new RandomPowerGenerator();
     }
 
-    RacingGame(final CarGroup carGroup, final PowerGenerator powerGenerator) {
-        this.carGroup = carGroup;
+    RacingGame(final String[] carNames, final PowerGenerator powerGenerator) {
+        this.carGroup = makeCarGroup(carNames);
         this.powerGenerator = powerGenerator;
     }
 
-    public void moveCars() {
-        carGroup.getAllCars().forEach(car -> {
-            final int power = powerGenerator.getNumber();
-            car.move(power);
-        });
+    public void move() {
+        carGroup.forEach(car -> car.move(powerGenerator.getNumber()));
+    }
+
+    public List<Car> getAllCars() {
+        return carGroup;
+    }
+
+    public List<Car> getWinners() {
+        final int positionMax = getPositionMax();
+
+        if (positionMax == 0) {
+            return List.of();
+        }
+
+        return carGroup.stream()
+                .filter(car -> car.isPositionedAt(positionMax))
+                .toList();
+    }
+
+    private List<Car> makeCarGroup(final String[] carNames) {
+        return Arrays.stream(carNames)
+                .map(Car::new)
+                .collect(Collectors.toList());
+    }
+
+    private int getPositionMax() {
+        return carGroup.stream()
+                .mapToInt(Car::getPosition)
+                .max()
+                .orElse(0);
     }
 }
